+++ content_creator.py
@@ -5,8 +5,9 @@ import json
 import glob
 import shutil
 import zipfile
+import io
 
-from package_creator import DIR_TO_PREFIX, merge_script_package_to_yml
+from package_creator import DIR_TO_PREFIX, merge_script_package_to_yml, write_yaml_with_docker
 
 CONTENT_DIRS = ['Integrations', 'Misc', 'Playbooks', 'Reports', 'Dashboards', 'Widgets', 'Scripts', 'IncidentTypes',
                 'Classifiers', 'Layouts', 'IncidentFields', 'IndicatorFields', 'Connections', 'Beta_Integrations']
@@ -14,11 +15,9 @@ CONTENT_DIRS = ['Integrations', 'Misc', 'Playbooks', 'Reports', 'Dashboards', 'W
 TEST_DIR = 'TestPlaybooks'
 
 # temp folder names
-BUNDLE_PRE = 'bundle_pre'
 BUNDLE_POST = 'bundle_post'
 BUNDLE_TEST = 'bundle_test'
 # zip files names (the extension will be added later - shutil demands file name without extension)
-ZIP_PRE = 'content_yml'
 ZIP_POST = 'content_new'
 ZIP_TEST = 'content_test'
 
@@ -63,7 +62,23 @@ def convert_incident_fields_to_array():
                 f.truncate()
 
 
-def copy_dir_yml(dir_name, version_num, bundle_pre, bundle_post, bundle_test):
+def copy_yaml_post(path, out_path, yml_info):
+    dirname = os.path.dirname(path)
+    if dirname in DIR_TO_PREFIX.keys() and not os.path.basename(path).startswith('playbook-'):
+        script_obj = yml_info
+        if dirname != 'Scripts':
+            script_obj = yml_info['script']
+        with io.open(path, mode='r', encoding='utf-8') as f:
+            yml_text = f.read()
+        out_map = write_yaml_with_docker(out_path, yml_text, yml_info, script_obj)
+        if len(out_map.keys()) > 1:
+            print(" - yaml generated multiple files: {}".format(out_map.keys()))
+        return
+    # not a script or integration file. Simply copy
+    shutil.copyfile(path, out_path)
+
+
+def copy_dir_yml(dir_name, bundle_post):
     scan_files = glob.glob(os.path.join(dir_name, '*.yml'))
     post_files = 0
     for path in scan_files:
@@ -71,36 +86,24 @@ def copy_dir_yml(dir_name, version_num, bundle_pre, bundle_post, bundle_test):
             yml_info = yaml.safe_load(f)
 
         ver = yml_info.get('fromversion', '0')
-        if ver == '' or is_ge_version(version_num, ver):
-            print ' - marked as post: %s (%s)' % (ver, path, )
-            shutil.copyfile(path, os.path.join(bundle_post, os.path.basename(path)))
-            shutil.copyfile(path, os.path.join(bundle_test, os.path.basename(path)))
-            post_files += 1
-        else:
-            # add the file to both bundles
-            print ' - marked as pre: %s (%s)' % (ver, path, )
-            shutil.copyfile(path, os.path.join(bundle_pre, os.path.basename(path)))
-            shutil.copyfile(path, os.path.join(bundle_post, os.path.basename(path)))
-
-    print ' - total post files: %d' % (post_files, )
+        print ' - processing: %s (%s)' % (ver, path, )
+        copy_yaml_post(path, os.path.join(bundle_post, os.path.basename(path)), yml_info)
+        post_files += 1
+    print ' - total files: %d' % (post_files, )
 
 
-def copy_dir_json(dir_name, version_num, bundle_pre, bundle_post, bundle_test):
+def copy_dir_json(dir_name, bundle_post):
     # handle *.json files
     scan_files = glob.glob(os.path.join(dir_name, '*.json'))
     for path in scan_files:
         dpath = os.path.basename(path)
-        shutil.copyfile(path, os.path.join(bundle_pre, os.path.basename(path)))
         # this part is a workaround because server doesn't support indicatorfield-*.json naming
-        shutil.copyfile(path, os.path.join(bundle_test, os.path.basename(path)))
         if dir_name == 'IndicatorFields':
             new_path = dpath.replace('incidentfield-', 'incidentfield-indicatorfield-')
             if os.path.isfile(new_path):
                 raise NameError('Failed while trying to create {}. File already exists.'.format(new_path))
             dpath = new_path
         shutil.copyfile(path, os.path.join(bundle_post, dpath))
-        shutil.copyfile(path, os.path.join(bundle_pre, dpath))
-        shutil.copyfile(path, os.path.join(bundle_test, dpath))
 
 
 def copy_dir_files(*args):
@@ -128,41 +131,35 @@ def copy_test_files(bundle_test):
 def main(circle_artifacts):
     print 'starting create content artifact ...'
 
-    # version that separate post bundle from pre bundle
-    # e.i. any yml with "fromversion" of <version_num> or more will be only on post bundle
-    version_num = "3.5"
-
     print 'creating dir for bundles ...'
-    for b in [BUNDLE_PRE, BUNDLE_POST, BUNDLE_TEST]:
+    for b in [BUNDLE_POST, BUNDLE_TEST]:
         os.mkdir(b)
-        add_tools_to_bundle(b)
+
+    add_tools_to_bundle(BUNDLE_POST)
 
     convert_incident_fields_to_array()
 
     for d in DIR_TO_PREFIX.keys():
         scanned_packages = glob.glob(os.path.join(d, '*/'))
         for package in scanned_packages:
-            merge_script_package_to_yml(package, d)
+            merge_script_package_to_yml(package, d, BUNDLE_POST)
 
     for d in CONTENT_DIRS:
         print 'copying dir %s to bundles ...' % (d,)
-        copy_dir_files(d, version_num, BUNDLE_PRE, BUNDLE_POST, BUNDLE_TEST)
+        copy_dir_files(d, BUNDLE_POST)
 
     copy_test_files(BUNDLE_TEST)
 
     print 'copying content descriptor to bundles'
-    for b in [BUNDLE_PRE, BUNDLE_POST, BUNDLE_TEST]:
+    for b in [BUNDLE_POST, BUNDLE_TEST]:
         shutil.copyfile('content-descriptor.json', os.path.join(b, 'content-descriptor.json'))
 
     print 'copying common server doc to bundles'
-    for b in [BUNDLE_PRE, BUNDLE_POST, BUNDLE_TEST]:
-        shutil.copyfile('./Documentation/doc-CommonServer.json', os.path.join(b, 'doc-CommonServer.json'))
+    shutil.copyfile('./Documentation/doc-CommonServer.json', os.path.join(BUNDLE_POST, 'doc-CommonServer.json'))
 
     print 'compressing bundles ...'
     shutil.make_archive(ZIP_POST, 'zip', BUNDLE_POST)
-    shutil.make_archive(ZIP_PRE, 'zip', BUNDLE_PRE)
     shutil.make_archive(ZIP_TEST, 'zip', BUNDLE_TEST)
-    shutil.copyfile(ZIP_PRE + '.zip', os.path.join(circle_artifacts, ZIP_PRE + '.zip'))
     shutil.copyfile(ZIP_POST + '.zip', os.path.join(circle_artifacts, ZIP_POST + '.zip'))
     shutil.copyfile(ZIP_TEST + '.zip', os.path.join(circle_artifacts, ZIP_TEST + '.zip'))
     shutil.copyfile("./Tests/id_set.json", os.path.join(circle_artifacts, "id_set.json"))