+++ Integrations/Kafka_V2/CHANGELOG.md
@@ -1,4 +1,16 @@
 ## [Unreleased]
+Updated the Docker image ***demisto/pykafka*** to version 1.0.0.3321 (requires Demisto 5.0).
+
+## [19.10.2] - 2019-10-29
+-
+
+## [19.10.1] - 2019-10-15
+  - Added partitions outputs to the ***kafka-print-topic*** command.
+  - Added the *Max number of messages to fetch* parameter.
+  - Added the *Use TLS for connection* parameter.
+  - Improved debug logging outputs.
+  - Improved fetch incidents implementation (breaks backward compatibility).
+  
 
 
 ## [19.9.1] - 2019-09-18